knowledge_store: rename service and locations modules
