+++ surfsense_backend/app/knowledge_store/__init__.py
@@ -1,13 +1,10 @@
-"""Git-native versioned storage for workspace knowledge.
-
-Public surface: build a :class:`KnowledgeStore` for a workspace and read/commit
-its history through it. Git is an implementation detail behind the facade.
-"""
+"""Git-native versioned storage for workspace knowledge."""
 
 from __future__ import annotations
 
 from app.knowledge_store.backends.base import StoredRevision
-from app.knowledge_store.service import KnowledgeStore, RevisionDraft
+from app.knowledge_store.revision_draft import RevisionDraft
+from app.knowledge_store.store import KnowledgeStore
 
 __all__ = [
     "KnowledgeStore",
